+++ package.json
@@ -23,7 +23,9 @@
     "mongoose": "^8.18.1",
     "morgan": "^1.10.1",
     "nodemailer": "^7.0.6",
-    "nodemon": "^3.1.10"
+    "nodemon": "^3.1.10",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^5.0.1"
   },
   "devDependencies": {
     "env-cmd": "^11.0.0"
