--- service/catway.js
@@ -1,7 +1,0 @@
-/**
- * logique des routes catway
- */
-
-exports.getById = (req, res) => {
-  // TODO : implémenter la logique pour récupérer une catway par ID
-}
